ref(getting-started-docs): Clean up docs load
