+++ static/app/components/onboarding/gettingStartedDoc/sdkDocumentation.tsx
@@ -6,89 +6,6 @@ import {PlatformKey} from 'sentry/data/platformCategories';
 import type {Organization, PlatformIntegration, Project, ProjectKey} from 'sentry/types';
 import {useApiQuery} from 'sentry/utils/queryClient';
 
-// Documents already migrated from sentry-docs to main sentry repository
-export const migratedDocs = [
-  'javascript-react',
-  'javascript-remix',
-  'javascript-angular',
-  'javascript-vue',
-  'javascript-gatsby',
-  'javascript-ember',
-  'javascript-svelte',
-  'javascript-sveltekit',
-  'javascript-nextjs',
-  'javascript',
-  'python-django',
-  'python',
-  'python-flask',
-  'python-wsgi',
-  'python-tryton',
-  'python-tornado',
-  'python-starlette',
-  'python-serverless',
-  'python-sanic',
-  'python-quart',
-  'python-pyramid',
-  'python-pylons',
-  'python-gcpfunctions',
-  'python-falcon',
-  'python-chalice',
-  'python-bottle',
-  'python-fastapi',
-  'python-asgi',
-  'python-aiohttp',
-  'python-awslambda',
-  'dotnet',
-  'dotnet-aspnetcore',
-  'dotnet-awslambda',
-  'dotnet-gcpfunctions',
-  'dotnet-maui',
-  'dotnet-uwp',
-  'dotnet-winforms',
-  'dotnet-wpf',
-  'dotnet-xamarin',
-  'dotnet-aspnet',
-  'react-native',
-  'java',
-  'java-spring-boot',
-  'java-logback',
-  'java-log4j2',
-  'java-spring',
-  'php',
-  'php-laravel',
-  'php-symfony',
-  'go',
-  'rust',
-  'minidump',
-  'native',
-  'native-qt',
-  'ruby',
-  'ruby-rails',
-  'ruby-rack',
-  'kotlin',
-  'node',
-  'node-awslambda',
-  'node-azurefunctions',
-  'node-connect',
-  'node-express',
-  'node-gcpfunctions',
-  'node-koa',
-  'node-serverlesscloud',
-  'electron',
-  'elixir',
-  'android',
-  'ionic',
-  'unity',
-  'capacitor',
-  'cordova',
-  'unreal',
-  'apple',
-  'apple-ios',
-  'apple-macos',
-  'flutter',
-  'dart',
-];
-
 type SdkDocumentationProps = {
   activeProductSelection: ProductSolution[];
   organization: Organization;
