+++ static/app/views/onboarding/setupDocs.tsx
@@ -3,26 +3,13 @@ import {browserHistory} from 'react-router';
 import styled from '@emotion/styled';
 import {motion} from 'framer-motion';
 
-import {loadDocs} from 'sentry/actionCreators/projects';
-import {Alert} from 'sentry/components/alert';
-import LoadingError from 'sentry/components/loadingError';
-import {DocumentationWrapper} from 'sentry/components/onboarding/documentationWrapper';
-import {
-  migratedDocs,
-  SdkDocumentation,
-} from 'sentry/components/onboarding/gettingStartedDoc/sdkDocumentation';
-import {MissingExampleWarning} from 'sentry/components/onboarding/missingExampleWarning';
+import {SdkDocumentation} from 'sentry/components/onboarding/gettingStartedDoc/sdkDocumentation';
 import {ProductSolution} from 'sentry/components/onboarding/productSelection';
-import {PlatformKey} from 'sentry/data/platformCategories';
 import platforms from 'sentry/data/platforms';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {Organization, Project} from 'sentry/types';
-import {OnboardingPlatformDoc} from 'sentry/types/onboarding';
 import {trackAnalytics} from 'sentry/utils/analytics';
-import getDynamicText from 'sentry/utils/getDynamicText';
 import {platformToIntegrationMap} from 'sentry/utils/integrationUtil';
-import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
 import SetupIntroduction from 'sentry/views/onboarding/components/setupIntroduction';
 import {SetupDocsLoader} from 'sentry/views/onboarding/setupDocsLoader';
@@ -31,72 +18,19 @@ import FirstEventFooter from './components/firstEventFooter';
 import IntegrationSetup from './integrationSetup';
 import {StepProps} from './types';
 
-function ProjectDocs(props: {
-  hasError: boolean;
-  onRetry: () => void;
-  organization: Organization;
-  platform: PlatformKey | null;
-  platformDocs: OnboardingPlatformDoc | null;
-  project: Project;
-}) {
-  const currentPlatform = props.platform ?? props.project?.platform ?? 'other';
-
-  return (
-    <Fragment>
-      <SetupIntroduction
-        stepHeaderText={t(
-          'Configure %s SDK',
-          platforms.find(p => p.id === currentPlatform)?.name ?? ''
-        )}
-        platform={currentPlatform}
-      />
-      {getDynamicText({
-        value: !props.hasError ? (
-          props.platformDocs !== null && (
-            <DocsWrapper key={props.platformDocs.html}>
-              <DocumentationWrapper
-                dangerouslySetInnerHTML={{__html: props.platformDocs.html}}
-              />
-              <MissingExampleWarning
-                platform={props.platform}
-                platformDocs={props.platformDocs}
-              />
-            </DocsWrapper>
-          )
-        ) : (
-          <LoadingError
-            message={t(
-              'Failed to load documentation for the %s platform.',
-              props.project?.platform
-            )}
-            onRetry={props.onRetry}
-          />
-        ),
-        fixed: (
-          <Alert type="warning">
-            Platform documentation is not rendered in for tests in CI
-          </Alert>
-        ),
-      })}
-    </Fragment>
-  );
-}
-
 function SetupDocs({location, recentCreatedProject: project}: StepProps) {
-  const api = useApi();
   const organization = useOrganization();
+
+  const [integrationUseManualSetup, setIntegrationUseManualSetup] = useState(false);
 
   const products = useMemo<ProductSolution[]>(
     () => (location.query.product ?? []) as ProductSolution[],
     [location.query.product]
   );
 
-  // SDK instrumentation
-  const [hasError, setHasError] = useState(false);
-  const [platformDocs, setPlatformDocs] = useState<OnboardingPlatformDoc | null>(null);
-  const [loadedPlatform, setLoadedPlatform] = useState<PlatformKey | null>(null);
+  const currentPlatformKey = project?.platform ?? 'other';
+  const currentPlatform = platforms.find(p => p.id === currentPlatformKey);
 
-  const currentPlatformKey = loadedPlatform ?? project?.platform ?? 'other';
   const [showLoaderOnboarding, setShowLoaderOnboarding] = useState(
     currentPlatformKey === 'javascript'
   );
@@ -104,14 +38,6 @@ function SetupDocs({location, recentCreatedProject: project}: StepProps) {
   useEffect(() => {
     setShowLoaderOnboarding(currentPlatformKey === 'javascript');
   }, [currentPlatformKey]);
-
-  const integrationSlug = project?.platform && platformToIntegrationMap[project.platform];
-  const [integrationUseManualSetup, setIntegrationUseManualSetup] = useState(false);
-
-  const showIntegrationOnboarding = integrationSlug && !integrationUseManualSetup;
-  const showDocsWithProductSelection =
-    currentPlatformKey.match('^javascript-([A-Za-z]+)$') ??
-    (showLoaderOnboarding === false && currentPlatformKey === 'javascript');
 
   const hideLoaderOnboarding = useCallback(() => {
     setShowLoaderOnboarding(false);
@@ -127,70 +53,13 @@ function SetupDocs({location, recentCreatedProject: project}: StepProps) {
     });
   }, [organization, currentPlatformKey, project?.id]);
 
-  const fetchData = useCallback(async () => {
-    // TODO: add better error handling logic
-    if (!project?.platform) {
-      return;
-    }
-
-    // this will be fetched in the DocWithProductSelection component
-    if (showDocsWithProductSelection) {
-      return;
-    }
-
-    // Show loader setup for base javascript platform
-    if (showLoaderOnboarding) {
-      return;
-    }
-
-    if (showIntegrationOnboarding) {
-      setLoadedPlatform(project.platform);
-      setPlatformDocs(null);
-      setHasError(false);
-      return;
-    }
-
-    // There are no docs for the other platform
-    if (project.platform === 'other') {
-      return;
-    }
-
-    try {
-      const loadedDocs = await loadDocs({
-        api,
-        orgSlug: organization.slug,
-        projectSlug: project.slug,
-        platform: project.platform as PlatformKey,
-      });
-      setPlatformDocs(loadedDocs);
-      setLoadedPlatform(project.platform);
-      setHasError(false);
-    } catch (error) {
-      setHasError(error);
-      throw error;
-    }
-  }, [
-    project?.slug,
-    project?.platform,
-    api,
-    organization.slug,
-    showDocsWithProductSelection,
-    showIntegrationOnboarding,
-    showLoaderOnboarding,
-  ]);
-
-  useEffect(() => {
-    fetchData();
-  }, [fetchData, location.query.product, project?.platform]);
-
-  if (!project) {
+  if (!project || !currentPlatform) {
     return null;
   }
 
-  const currentPlatform = platforms.find(p => p.id === currentPlatformKey);
   const platformName = currentPlatform?.name ?? '';
-  const loadLocalSdkDocumentation =
-    currentPlatform && migratedDocs.includes(currentPlatformKey);
+  const integrationSlug = project?.platform && platformToIntegrationMap[project.platform];
+  const showIntegrationOnboarding = integrationSlug && !integrationUseManualSetup;
 
   return (
     <Fragment>
@@ -204,51 +73,34 @@ function SetupDocs({location, recentCreatedProject: project}: StepProps) {
                 setIntegrationUseManualSetup(true);
               }}
             />
-          ) : showLoaderOnboarding ? (
-            <Fragment>
-              <SetupIntroduction
-                stepHeaderText={t(
-                  'Configure %s SDK',
-                  platforms.find(p => p.id === currentPlatformKey)?.name ?? ''
-                )}
-                platform={currentPlatformKey}
-              />
-              <SetupDocsLoader
-                organization={organization}
-                project={project}
-                location={location}
-                platform={loadedPlatform}
-                close={hideLoaderOnboarding}
-              />
-            </Fragment>
-          ) : loadLocalSdkDocumentation ? (
+          ) : (
             <Fragment>
               <SetupIntroduction
                 stepHeaderText={t('Configure %s SDK', platformName)}
                 platform={currentPlatformKey}
               />
-              <SdkDocumentation
-                platform={currentPlatform}
-                organization={organization}
-                projectSlug={project.slug}
-                projectId={project.id}
-                activeProductSelection={products}
-                newOrg
-              />
+              {showLoaderOnboarding ? (
+                <SetupDocsLoader
+                  organization={organization}
+                  project={project}
+                  location={location}
+                  platform={currentPlatform.id}
+                  close={hideLoaderOnboarding}
+                />
+              ) : (
+                <SdkDocumentation
+                  platform={currentPlatform}
+                  organization={organization}
+                  projectSlug={project.slug}
+                  projectId={project.id}
+                  activeProductSelection={products}
+                  newOrg
+                />
+              )}
             </Fragment>
-          ) : (
-            <ProjectDocs
-              platform={loadedPlatform}
-              project={project}
-              hasError={hasError}
-              platformDocs={platformDocs}
-              onRetry={fetchData}
-              organization={organization}
-            />
           )}
         </MainContent>
       </Wrapper>
-
       <FirstEventFooter
         project={project}
         organization={organization}
