+++ static/app/views/projectInstall/platform.tsx
@@ -3,48 +3,34 @@ import {RouteComponentProps} from 'react-router';
 import styled from '@emotion/styled';
 import omit from 'lodash/omit';
 
-import {loadDocs, removeProject} from 'sentry/actionCreators/projects';
 import Feature from 'sentry/components/acl/feature';
 import {Alert} from 'sentry/components/alert';
 import {Button} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
-import Confirm from 'sentry/components/confirm';
 import NotFound from 'sentry/components/errors/notFound';
 import HookOrDefault from 'sentry/components/hookOrDefault';
-import ExternalLink from 'sentry/components/links/externalLink';
-import LoadingError from 'sentry/components/loadingError';
-import LoadingIndicator from 'sentry/components/loadingIndicator';
-import {DocumentationWrapper} from 'sentry/components/onboarding/documentationWrapper';
-import {
-  migratedDocs,
-  SdkDocumentation,
-} from 'sentry/components/onboarding/gettingStartedDoc/sdkDocumentation';
+import {SdkDocumentation} from 'sentry/components/onboarding/gettingStartedDoc/sdkDocumentation';
 import {ProductSolution} from 'sentry/components/onboarding/productSelection';
-import {useRecentCreatedProject} from 'sentry/components/onboarding/useRecentCreatedProject';
-import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {
   performance as performancePlatforms,
   Platform,
   PlatformKey,
 } from 'sentry/data/platformCategories';
 import platforms from 'sentry/data/platforms';
-import {IconChevron} from 'sentry/icons';
-import {t, tct} from 'sentry/locale';
+import {t} from 'sentry/locale';
 import ConfigStore from 'sentry/stores/configStore';
 import {space} from 'sentry/styles/space';
-import {OnboardingSelectedSDK, Organization, Project} from 'sentry/types';
+import {OnboardingSelectedSDK} from 'sentry/types';
 import {IssueAlertRule} from 'sentry/types/alerts';
 import {trackAnalytics} from 'sentry/utils/analytics';
-import {handleXhrErrorResponse} from 'sentry/utils/handleXhrErrorResponse';
 import {useApiQuery} from 'sentry/utils/queryClient';
-import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjects from 'sentry/utils/useProjects';
-import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import {SetupDocsLoader} from 'sentry/views/onboarding/setupDocsLoader';
 import {GettingStartedWithProjectContext} from 'sentry/views/projects/gettingStartedWithProjectContext';
 
 import {OtherPlatformsInfo} from './otherPlatformsInfo';
+import {PlatformDocHeader} from './platformDocHeader';
 
 const ProductUnavailableCTAHook = HookOrDefault({
   hookName: 'component:product-unavailable-cta',
@@ -52,77 +38,8 @@ const ProductUnavailableCTAHook = HookOrDefault({
 
 type Props = RouteComponentProps<{projectId: string}, {}>;
 
-export function SetUpGeneralSdkDoc({
-  organization,
-  projectSlug,
-  platform,
-}: {
-  organization: Organization;
-  platform: Platform;
-  projectSlug: Project['slug'];
-}) {
-  const api = useApi();
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(false);
-  const [html, setHtml] = useState('');
-
-  const fetchDocs = useCallback(async () => {
-    setLoading(true);
-
-    try {
-      const {html: reponse} = await loadDocs({
-        api,
-        orgSlug: organization.slug,
-        projectSlug,
-        platform: platform.key as PlatformKey,
-      });
-      setHtml(reponse);
-      window.scrollTo(0, 0);
-    } catch (err) {
-      setError(err);
-    }
-
-    setLoading(false);
-  }, [api, organization.slug, projectSlug, platform.key]);
-
-  useEffect(() => {
-    fetchDocs();
-  }, [fetchDocs]);
-
-  return (
-    <div>
-      <Alert type="info" showIcon>
-        {tct(
-          `
-           This is a quick getting started guide. For in-depth instructions
-           on integrating Sentry with [platform], view
-           [docLink:our complete documentation].`,
-          {
-            platform: platform.name,
-            docLink: <ExternalLink href={platform.link ?? undefined} />,
-          }
-        )}
-      </Alert>
-      {loading ? (
-        <LoadingIndicator />
-      ) : error ? (
-        <LoadingError onRetry={fetchDocs} />
-      ) : (
-        <Fragment>
-          <SentryDocumentTitle
-            title={`${t('Configure')} ${platform.name}`}
-            projectSlug={projectSlug}
-          />
-          <DocumentationWrapper dangerouslySetInnerHTML={{__html: html}} />
-        </Fragment>
-      )}
-    </div>
-  );
-}
-
-export function ProjectInstallPlatform({location, params, router}: Props) {
+export function ProjectInstallPlatform({location, params}: Props) {
   const organization = useOrganization();
-  const api = useApi();
   const gettingStartedWithProjectContext = useContext(GettingStartedWithProjectContext);
 
   const isSelfHosted = ConfigStore.get('isSelfHosted');
@@ -201,35 +118,12 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
     projectAlertRulesIsError,
   ]);
 
-  const projectDeletionOnBackClick = !!organization?.features.includes(
-    'onboarding-project-deletion-on-back-click'
-  );
-
   // This is a feature flag that is currently only enabled for a subset of internal users until the feature is fully implemented,
   // but the purpose of the feature is to make the product selection feature in documents available to all users
   // and guide them to upgrade to a plan if one of the products is not available on their current plan.
   const gettingStartedDocWithProductSelection = !!organization?.features.includes(
     'getting-started-doc-with-product-selection'
   );
-
-  const recentCreatedProject = useRecentCreatedProject({
-    orgSlug: organization.slug,
-    projectSlug: project?.slug,
-  });
-
-  const shallProjectBeDeleted =
-    projectDeletionOnBackClick &&
-    recentCreatedProject &&
-    // if the project has received a first error, we don't delete it
-    recentCreatedProject.firstError === false &&
-    // if the project has received a first transaction, we don't delete it
-    recentCreatedProject.firstTransaction === false &&
-    // if the project has replays, we don't delete it
-    recentCreatedProject.hasReplays === false &&
-    // if the project has sessions, we don't delete it
-    recentCreatedProject.hasSessions === false &&
-    // if the project is older than one hour, we don't delete it
-    recentCreatedProject.olderThanOneHour === false;
 
   const platformIntegration = platforms.find(p => p.id === currentPlatformKey);
   const platform: Platform = {
@@ -238,49 +132,6 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
     name: platformIntegration?.name,
     link: platformIntegration?.link,
   };
-
-  const handleGoBack = useCallback(async () => {
-    if (!recentCreatedProject) {
-      return;
-    }
-
-    trackAnalytics('project_creation.back_button_clicked', {
-      organization,
-    });
-
-    if (shallProjectBeDeleted) {
-      trackAnalytics('project_creation.data_removal_modal_confirm_button_clicked', {
-        organization,
-        platform: recentCreatedProject.slug,
-        project_id: recentCreatedProject.id,
-      });
-
-      try {
-        await removeProject({
-          api,
-          orgSlug: organization.slug,
-          projectSlug: recentCreatedProject.slug,
-          origin: 'getting_started',
-        });
-
-        trackAnalytics('project_creation.data_removed', {
-          organization,
-          date_created: recentCreatedProject.dateCreated,
-          platform: recentCreatedProject.slug,
-          project_id: recentCreatedProject.id,
-        });
-      } catch (error) {
-        handleXhrErrorResponse('Unable to delete project in project creation', error);
-        // we don't give the user any feedback regarding this error as this shall be silent
-      }
-    }
-
-    router.replace(
-      normalizeUrl(
-        `/organizations/${organization.slug}/projects/new/?referrer=getting-started&project=${recentCreatedProject.id}`
-      )
-    );
-  }, [api, recentCreatedProject, organization, shallProjectBeDeleted, router]);
 
   const hideLoaderOnboarding = useCallback(() => {
     setShowLoaderOnboarding(false);
@@ -296,7 +147,7 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
     });
   }, [organization, currentPlatform, project?.id]);
 
-  if (!project) {
+  if (!project || !currentPlatform) {
     return null;
   }
 
@@ -308,7 +159,6 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
   const performanceOverviewLink = `/organizations/${organization.slug}/performance/`;
   const showPerformancePrompt = performancePlatforms.includes(platform.id as PlatformKey);
   const isGettingStarted = window.location.href.indexOf('getting-started') > 0;
-
   const showDocsWithProductSelection =
     gettingStartedDocWithProductSelection &&
     (platform.key === 'javascript' || !!platform.key.match('^javascript-([A-Za-z]+)$'));
@@ -318,82 +168,29 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
       {!isSelfHosted && showDocsWithProductSelection && (
         <ProductUnavailableCTAHook organization={organization} />
       )}
-      <StyledPageHeader>
-        <h2>{t('Configure %(platform)s SDK', {platform: platform.name ?? 'other'})}</h2>
-        <ButtonBar gap={1}>
-          <Confirm
-            bypass={!shallProjectBeDeleted}
-            message={t(
-              "Hey, just a heads up - we haven't received any data for this SDK yet and by going back all changes will be discarded. Are you sure you want to head back?"
-            )}
-            priority="danger"
-            confirmText={t("Yes I'm sure")}
-            onConfirm={handleGoBack}
-            onClose={() => {
-              if (!recentCreatedProject) {
-                return;
-              }
-              trackAnalytics('project_creation.data_removal_modal_dismissed', {
-                organization,
-                platform: recentCreatedProject.slug,
-                project_id: recentCreatedProject.id,
-              });
-            }}
-            onRender={() => {
-              if (!recentCreatedProject) {
-                return;
-              }
-              trackAnalytics('project_creation.data_removal_modal_rendered', {
-                organization,
-                platform: recentCreatedProject.slug,
-                project_id: recentCreatedProject.id,
-              });
-            }}
-          >
-            <Button icon={<IconChevron direction="left" size="sm" />} size="sm">
-              {t('Back to Platform Selection')}
-            </Button>
-          </Confirm>
-          {platform.key !== 'other' && (
-            <Button size="sm" href={platform.link ?? undefined} external>
-              {t('Full Documentation')}
-            </Button>
-          )}
-        </ButtonBar>
-      </StyledPageHeader>
+      <PlatformDocHeader projectSlug={project.slug} platform={platform} />
       {platform.key === 'other' ? (
         <OtherPlatformsInfo
           projectSlug={project.slug}
           platform={platform.name ?? 'other'}
         />
+      ) : showLoaderOnboarding ? (
+        <SetupDocsLoader
+          organization={organization}
+          project={project}
+          location={location}
+          platform={currentPlatform.id}
+          close={hideLoaderOnboarding}
+        />
       ) : (
-        <Fragment>
-          {currentPlatform && showLoaderOnboarding ? (
-            <SetupDocsLoader
-              organization={organization}
-              project={project}
-              location={location}
-              platform={currentPlatform.id}
-              close={hideLoaderOnboarding}
-            />
-          ) : currentPlatform && migratedDocs.includes(currentPlatformKey) ? (
-            <SdkDocumentation
-              platform={currentPlatform}
-              organization={organization}
-              projectSlug={project.slug}
-              projectId={project.id}
-              activeProductSelection={products}
-            />
-          ) : (
-            <SetUpGeneralSdkDoc
-              organization={organization}
-              projectSlug={project.slug}
-              platform={platform}
-            />
-          )}
-        </Fragment>
+        <SdkDocumentation
+          platform={currentPlatform}
+          organization={organization}
+          projectSlug={project.slug}
+          projectId={project.id}
+          activeProductSelection={products}
+        />
       )}
-
       <div>
         {isGettingStarted && showPerformancePrompt && (
           <Feature
@@ -414,7 +211,6 @@ export function ProjectInstallPlatform({location, params, router}: Props) {
             }}
           </Feature>
         )}
-
         <StyledButtonBar gap={1}>
           <Button
             priority="primary"
@@ -450,25 +246,6 @@ const StyledButtonBar = styled(ButtonBar)`
     width: auto;
     grid-row-gap: ${space(1)};
     grid-auto-flow: row;
-  }
-`;
-
-const StyledPageHeader = styled('div')`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: ${space(3)};
-
-  h2 {
-    margin: 0;
-  }
-
-  @media (max-width: ${p => p.theme.breakpoints.small}) {
-    flex-direction: column;
-    align-items: flex-start;
-
-    h2 {
-      margin-bottom: ${space(2)};
-    }
   }
 `;
